transSum-server/pyproject.toml
[project]
name = "transsum"
version = "0.1.0"
description = "Summarize & translate documents via Ollama or Anthropic, with MCP integration."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "click>=8.1",
    "httpx>=0.27",
    "mcp[cli]>=1.26.0",
    "pydantic-settings>=2.0",
    "pypdf>=4.0",
    "python-dotenv>=1.2.1",
    "rich>=13.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]

[project.scripts]
transsum = "transsum.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

# ── Ruff (linting & formatting — pairs well with uv) ───────────────────
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

# ── Pytest ──────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"